Add useCheckout hook in order to not duplicate 'go to checkout' logic
