+++ react/CheckoutButton.tsx
@@ -3,6 +3,7 @@ import { FormattedMessage } from 'react-intl'
 import { Button } from 'vtex.styleguide'
 import { useCssHandles } from 'vtex.css-handles'
 import { useCheckoutURL } from 'vtex.checkout-resources/Utils'
+import useCheckout from './modules/checkoutHook'
 
 interface Props {
   finishShoppingButtonLink: string
@@ -13,12 +14,13 @@ const CSS_HANDLES = ['minicartCheckoutButton'] as const
 const CheckoutButton: FC<Props> = ({ finishShoppingButtonLink }) => {
   const { url: checkoutUrl } = useCheckoutURL()
   const handles = useCssHandles(CSS_HANDLES)
+  const goToCheckout = useCheckout()
 
   return (
     <div className={`${handles.minicartCheckoutButton} mv3`}>
       <Button
         id="proceed-to-checkout"
-        href={finishShoppingButtonLink || checkoutUrl}
+        onClick={() => goToCheckout(finishShoppingButtonLink || checkoutUrl)}
         variation="primary"
         block
       >
